Add AddressAnalyses

+++ db/migrate/20211118152519_create_analysis_results.rb
@@ -4,7 +4,7 @@ class CreateAnalysisResults < ActiveRecord::Migration[6.1]
     enable_extension 'citext'
     create_table :analysis_results do |t|
       t.citext :address, null: false
-      t.integer :risk_confidence, null: false
+      t.decimal :risk_confidence, null: false
       t.integer :risk_level, null: false
       t.jsonb :raw_response, null: false
