+++ db/migrate/20211119063003_create_address_analyses.rb
@@ -3,7 +3,7 @@ class CreateAddressAnalyses < ActiveRecord::Migration[6.1]
     create_table :address_analyses do |t|
       t.citext :address, null: false
       t.integer :risk_level, null: false
-      t.integer :risk_confidence, null: false
+      t.decimal :risk_confidence, null: false
       t.references :analysis_result, null: false, foreign_key: true
 
       t.timestamps
